+++ NEWS.md
@@ -1,3 +1,13 @@
+# AlphaSimR 1.6.0
+
+*changed misc slot in Pop class from a list organised as ind x nodes to to a list organised as nodes x ind (this simplified code and increased speed)
+
+*removed `setMisc` and `getMisc`
+
+*added `length` method for Pop class that returns number of individuals (like `nInd`)
+
+*added `length` method for MultiPop class that returns number of populations
+
 # AlphaSimR 1.5.3
 
 *fixed bug in `SimParam$restrSegSites` with excluding sites at end of chromosome